Make http client optionally configurable
