+++ options.go
@@ -1,0 +1,17 @@
+package sensonet
+
+import "net/http"
+
+type Option func(*Connection)
+
+func WithLogger(logger Logger) Option {
+	return func(c *Connection) {
+		c.logger = logger
+	}
+}
+
+func WithHttpClient(client *http.Client) Option {
+	return func(c *Connection) {
+		c.client = client
+	}
+}
